+++ CHANGELOG.md
@@ -21,12 +21,16 @@ to [Semantic Versioning](https://semver.org).
 
 ### Fixed
 
+### Security
+
+## [0.17.1] - 2022-12-04
+
+### Fixed
+
 - #23 - NullPointerException while open a project which have some invalid metadata.
 - #25 - IllegalAccessError while request a document for value of a Property Group.
 - #31 - IllegalAccessError while request document for a misplaced property value.
 - #43 - NullPointerException while reporting an empty error.
-
-### Security
 
 ## [0.17.0] - 2022-11-07